Move future sort to appointment utils
